DEVONthink3 & mutliple type support
The application name for the new version of DEVONthink has changed, so I’ve changed the application selection to match. If you’re using DEVONthink 2 you’ll need to change it back.

The only feature change is to add support for multiple types in the functions which checks the selection is of the right file type. The content_type environment variable is set in the “Args and Vars” utility and uses a semicolon sepeator, i.e.,: “PDF+Text;PDF Document;markdown” is use for the “dp” keyword to launch the system default application. The prior version could only have 1 content type set, which was a bit annoying.

I’ve added a decompiled version of the script here too if you want to see what it does before downloading.